requirements.txt
psycopg2-binary==2.9.6
PyYAML==6.0
schedule==1.2.0
SQLAlchemy==2.0.17

